requirements.txt
kornia==0.6.8
numpy==1.23.5
Pillow==9.4.0
scipy==1.9.3
googledrivedownloader
requests
